backend/requirements.txt
fastapi
uvicorn
transformers
torch
pillow
python-multipart
openai-whisper
sqlmodel
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
email-validator
opencv-python
